Move `controllers/notification.js` into `middleware/`
because it's not actually a controller, it's middleware
